Added a diagramType property to NodeAction so that the actions can be hidden and shown based on the diagram type.

+++ src/nodefactory.cpp
@@ -9,7 +9,7 @@ vector<BaseNode*> NodeFactory::prototypes;
 NodeFactory::~NodeFactory()
 {
     //delete the pointers
-    for (int i=0; i<prototypes.size(); i++) {
+    for (int i=0; i<(int)prototypes.size(); i++) {
         delete prototypes.at(i);
     }
 }
